Move it to a less general namespace
