+++ config-model/src/test/java/com/yahoo/vespa/model/admin/AdminTestCase.java
@@ -17,7 +17,7 @@ import com.yahoo.config.provision.Zone;
 import com.yahoo.container.StatisticsConfig;
 import com.yahoo.container.jdisc.config.HealthMonitorConfig;
 import com.yahoo.net.HostName;
-import com.yahoo.vespa.config.StateserverConfig;
+import com.yahoo.vespa.config.core.StateserverConfig;
 import com.yahoo.vespa.model.VespaModel;
 import com.yahoo.vespa.model.container.ContainerCluster;
 import com.yahoo.vespa.model.container.component.Component;
