+++ configdefinitions/src/main/java/com/yahoo/vespa/config/core/package-info.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 @ExportPackage
-package com.yahoo.vespa.config;
+package com.yahoo.vespa.config.core;
 
 import com.yahoo.osgi.annotation.ExportPackage;
