+++ configdefinitions/src/vespa/CMakeLists.txt
@@ -75,4 +75,4 @@ install_config_definition(zookeepers.def cloud.config.zookeepers.def)
 vespa_generate_config(configdefinitions bucketspaces.def)
 install_config_definition(bucketspaces.def vespa.config.content.core.bucketspaces.def)
 vespa_generate_config(configdefinitions stateserver.def)
-install_config_definition(stateserver.def vespa.config.stateserver.def)
+install_config_definition(stateserver.def vespa.config.core.stateserver.def)
