+++ configdefinitions/src/vespa/stateserver.def
@@ -1,5 +1,5 @@
 # Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-namespace=vespa.config
+namespace=vespa.config.core
 
 # Port used for serving the state/vN api
 httpport int default = 0
